+++ extensions/iceworks-project-creator/src/services/index.ts
@@ -1,5 +1,6 @@
 import * as project from '@iceworks/project-service';
 import * as material from '@iceworks/material-service';
 import * as common from '@iceworks/common-service';
+import * as pegasus from '@iceworks/pegasus-service';
 
-export default { project, common, material };
+export default { project, common, material, pegasus };