+++ packages/generate-material/src/ejsRenderDir.ts
@@ -2,11 +2,20 @@ import * as path from 'path';
 import * as glob from 'glob';
 import * as ejs from 'ejs';
 import * as fse from 'fs-extra';
+<<<<<<< HEAD
 
 export default async function (dir: string, options: any): Promise<void> {
   return new Promise((resolve, reject) => {
     glob(
       '**/*.ejs',
+=======
+import { ITemplateOptions } from './index';
+
+export default async function (dir: string, options: ITemplateOptions): Promise<void> {
+  return new Promise((resolve, reject) => {
+    glob(
+      '**/*.?(_)ejs',
+>>>>>>> origin/master
       {
         cwd: dir,
         nodir: true,
@@ -37,6 +46,7 @@ export default async function (dir: string, options: any): Promise<void> {
 
 function renderFile(filepath: string, options: any): Promise<string> {
   return new Promise((resolve, reject) => {
+<<<<<<< HEAD
     ejs.renderFile(filepath, options, (err, result) => {
       if (err) {
         return reject(err);
@@ -46,5 +56,21 @@ function renderFile(filepath: string, options: any): Promise<string> {
       fse.writeFileSync(filepath.replace(/\.ejs$/, ''), result);
       resolve();
     });
+=======
+    if (filepath.endsWith('_ejs')) {
+      const templateFilePath = filepath.replace(/_ejs$/, 'ejs');
+      fse.renameSync(filepath, templateFilePath);
+      resolve();
+    } else {
+      ejs.renderFile(filepath, options, (err, result) => {
+        if (err) {
+          return reject(err);
+        }
+        fse.removeSync(filepath);
+        fse.writeFileSync(filepath.replace(/\.ejs$/, ''), result);
+        resolve();
+      });
+    }
+>>>>>>> origin/master
   });
 }