+++ extensions/iceworks-config-helper/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+<<<<<<< HEAD
+=======
+## 0.3.2
+
+- docs: add more badges.
+
+>>>>>>> origin/master
 ## 0.3.1
 
 - feat: support all json files.